LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-volume-slider-visibility-expected.txt

Testing the MacOSInlineMediaControls volume slider visibility.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Default state
PASS volumeSliderContainer.parent is null

Mouse enters the mute button
PASS volumeSliderContainer.parent is mediaControls

Mouse leaves the mute button
PASS volumeSliderContainer.parent is null

Mouse enters the mute button, then leaves it but enters the volume slider container
PASS volumeSliderContainer.parent is mediaControls

Mouse leaves the volume slider container
PASS volumeSliderContainer.parent is null

PASS successfullyParsed is true

TEST COMPLETE
